Ratings Finished
Added ratings logic and basic UI to finished page. Ready for styling by @kally (was not sure how she wants to do this, so just used a package to make it work).

Test:
Go to recipe finished page for a recipe, press a star number, press finished, and it adds rating in firebase under relevant recipes and updates the total rating for the recipe in test_recipes as well.

NOTE:
Right now, a user could submit multiple ratings for the same recipe.